+++ GetResponseIntegration/Block/Settings.php
@@ -43,99 +43,4 @@ class Settings extends Template
     {
         return AccountFactory::createFromArray($this->repository->getAccountInfo());
     }
-
-    /**
-     * @return bool
-     */
-    public function getLastPostedApiKey()
-    {
-        /** @var Http $request */
-        $request = $this->getRequest();
-        $data = $request->getPostValue();
-        if (!empty($data)) {
-            if (isset($data['getresponse_api_key'])) {
-                return $data['getresponse_api_key'];
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * @return string
-     */
-    public function getHiddenApiKey()
-    {
-        $settings = $this->repository->getConnectionSettings();
-
-        if (empty($settings)) {
-            return '';
-        }
-
-        $settings = ConnectionSettingsFactory::createFromArray($settings);
-
-        if (empty($settings->getApiKey())) {
-            return '';
-        }
-
-        return strlen($settings->getApiKey()) > 0 ? str_repeat(
-            "*",
-            strlen($settings->getApiKey()) - 6
-        ) . substr($settings->getApiKey(), -6) : '';
-    }
-
-    /**
-     * @return int
-     */
-    public function getLastPostedApiAccount()
-    {
-        /** @var Http $request */
-        $request = $this->getRequest();
-        $data = $request->getPostValue();
-        if (!empty($data['getresponse_360_account']) && 1 == $data['getresponse_360_account']) {
-            return $data['getresponse_360_account'];
-        }
-
-        return 0;
-    }
-
-    /**
-     * @return bool
-     */
-    public function getLastPostedApiUrl()
-    {
-        /** @var Http $request */
-        $request = $this->getRequest();
-        $data = $request->getPostValue();
-        if (!empty($data['getresponse_api_url'])) {
-            return $data['getresponse_api_url'];
-        }
-
-        return false;
-    }
-
-    /**
-     * @return bool
-     */
-    public function getLastPostedApiDomain()
-    {
-        /** @var Http $request */
-        $request = $this->getRequest();
-        $data = $request->getPostValue();
-        if (!empty($data['getresponse_api_domain'])) {
-            return $data['getresponse_api_domain'];
-        }
-
-        return false;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isConnectedToGetResponse()
-    {
-        $settings = $this->repository->getConnectionSettings();
-
-        return !empty($settings['apiKey']);
-    }
 }
