+++ GetResponseIntegration/Block/WebEventTracking.php
@@ -20,10 +20,7 @@ class WebEventTracking extends Template
      * @param Context $context
      * @param Repository $repository
      */
-    public function __construct(
-        Context $context,
-        Repository $repository
-    ) {
+    public function __construct(Context $context, Repository $repository) {
         parent::__construct($context);
         $this->repository = $repository;
     }
