+++ GetResponseIntegration/Controller/Adminhtml/Rules/Create.php
@@ -52,7 +52,6 @@ class Create extends Action
      * Dispatch request
      *
      * @return \Magento\Framework\Controller\ResultInterface|ResponseInterface
-     * @throws \Magento\Framework\Exception\NotFoundException
      */
     public function execute()
     {
