+++ GetResponseIntegration/view/adminhtml/layout/getresponseintegration_account_delete.xml
@@ -8,7 +8,7 @@
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="admin-1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <body>
         <referenceContainer name="content">
-            <block name="getresponse.settings.form" class="GetResponse\GetResponseIntegration\Block\Settings" template="settings.phtml" />
+            <block name="getresponse.account" class="GetResponse\GetResponseIntegration\Block\Account" template="account.phtml" />
         </referenceContainer>
     </body>
 </page>
