+++ GetResponseIntegration/view/adminhtml/templates/create_list.phtml
@@ -40,7 +40,7 @@ $confirmation_bodies = $block->getSubscriptionConfirmationsBody();
                             <span>List name</span>
                         </label>
                         <div class="admin__field-control control">
-                            <input name="campaign_name" id="campaign_name" value="<?php echo $block->getLastPostedApiKey(); ?>" title="API Key" class="basic_behavior input-text input-text admin__control-text required-entry _required" type="text" aria-required="true">
+                            <input name="campaign_name" id="campaign_name" value="" title="List name" class="basic_behavior input-text input-text admin__control-text required-entry _required" type="text" aria-required="true">
                             <div class="validation-message"></div>
                         </div>
                     </div>
